Added validation; (#2)
* Added validation;

* Added init controller to add some users;
Added Swagger;

+++ src/main/java/com/example/cooking/model/dto/response/RecipeResponseDto.java
@@ -6,14 +6,13 @@ import lombok.Data;
 
 @Data
 public class RecipeResponseDto {
-
     private long id;
     private String name;
     private String description;
     //TODO: File photo
     private List<Long> dishTypeIds;
     private List<Long> mealTimeIds;
-    private Map<Long, Integer> ingredients;
+    private Map<Long, Long> ingredients;
     private List<Long> portionIds;
     private int preparationTime;
 }
